requirements.txt
# Основные зависимости Flask
flask==2.0.1
flask-cors==3.0.10

# Работа с базой данных (PostgreSQL)
psycopg2-binary==2.9.1

# Аутентификация и токены
pyjwt==2.3.0
bcrypt==3.2.0

# Логирование (встроено в Python, но могут потребоваться дополнительные обработчики)
# (Не требует установки)

# Дополнительные утилиты
python-dotenv==0.19.0  # Для загрузки .env-файлов
pyyaml==6.0            # Если конфиги в формате YAML
ldap3==2.9.1           # Если используется LDAP-аутентификация
cryptography==36.0.1   # Для дополнительной криптографии (если JWT с RS256)

# Для production-развертывания (опционально)
gunicorn==20.1.0